requirements.txt
api==0.0.7
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
nose==1.3.7
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
